Remove unused resources helpers, extract the deprecated ones
